+++ src/DesktopGrass.Native/src/Renderer.h
@@ -111,6 +111,7 @@ private:
     ComPtr<ID2D1SolidColorBrush>           snowLayerHighlightBrush_;
     ComPtr<ID2D1SolidColorBrush>           driftBaseBrush_;
     ComPtr<ID2D1SolidColorBrush>           driftHiliteBrush_;
+    ComPtr<ID2D1SolidColorBrush>           snowBankShadowBrush_;
     ComPtr<ID2D1SolidColorBrush>           pineBrush_;
     ComPtr<ID2D1SolidColorBrush>           pineShadowBrush_;
     ComPtr<ID2D1SolidColorBrush>           pineHighlightBrush_;